pyproject.toml
[project]
name = "stackerbee-teenstogather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "amqp==5.3.1",
    "anyio==4.6.0",
    "asgiref==3.8.1",
    "async-timeout==5.0.1",
    "billiard==4.2.1",
    "celery==5.4.0",
    "certifi==2024.8.30",
    "channels==4.1.0",
    "charset-normalizer==3.4.0",
    "click-didyoumean==0.3.1",
    "click==8.1.7",
    "sync>=1.0.0",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "django==5.1",
    "django-cors-headers==4.4.0",
    "django-filter==24.3",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.3.1",
    "drf-yasg==1.21.7",
    "exceptiongroup==1.2.2",
    "faker==30.0.0",
    "h11==0.14.0",
    "httptools==0.6.1",
    "idna==3.10",
    "inflection==0.5.1",
    "kombu==5.4.2",
    "markdown==3.7",
    "packaging==24.1",
    "pillow==10.0.1",
    "prompt-toolkit==3.0.48",
    "psycopg2==2.9.9",
    "psycopg2-binary==2.9.10",
    "pyjwt==2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "pyyaml==6.0.2",
    "redis==5.2.0",
    "requests==2.32.3",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.1",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
    "urllib3==2.2.3",
    "uvicorn==0.31.1",
    "vine==5.1.0",
    "watchfiles==0.24.0",
    "wcwidth==0.2.13",
    "websockets==13.1",
]
